fixed: session for solid
